Add config file for 0322 dataset with branch entropy and reuse distance

+++ DP/inst2mmap.py
@@ -39,7 +39,8 @@ for i in range(len(args.fname)):
         continue
 
       try:
-        vals = [int(val) for val in line.split()]
+        #vals = [int(val) for val in line.split()]
+        vals = [float(val) for val in line.split()]
         assert len(vals) == inst_length
       except:
         print("Bad line:", len(vals), vals, flush=True)
